+++ NEWS.md
@@ -1,3 +1,13 @@
+# cyclestreets 1.0.3 (2024-12)
+
+* Increased timeout time (#91)
+* Use retry
+* Added Wang Zhao to DESCRIPTION file
+
+# cyclestreets 1.0.2 (2024-09)
+
+* Minor update to fix an issue with the `batch()` function (#91)
+
 # cyclestreets 1.0.1
 
 * Bug fix: issue with batch identified and fix (#80) thanks to @mem48